test: refactor internal error reports to new scenario
