+++ features/fixtures/mazerunner/cxx-scenarios-bugsnag/detekt-baseline.xml
@@ -4,6 +4,5 @@
   <CurrentIssues>
     <ID>MagicNumber:CXXExceptionSmokeScenario.kt$CXXExceptionSmokeScenario$500</ID>
     <ID>MagicNumber:CXXExceptionSmokeScenario.kt$CXXExceptionSmokeScenario$999</ID>
-    <ID>MagicNumber:CXXNaughtyStringsScenario.kt$CXXNaughtyStringsScenario$200</ID>
   </CurrentIssues>
 </SmellBaseline>
