+++ features/fixtures/mazerunner/cxx-scenarios/src/main/java/com/bugsnag/android/mazerunner/scenarios/CXXPausedSessionScenario.kt
@@ -3,6 +3,8 @@ package com.bugsnag.android.mazerunner.scenarios
 import android.content.Context
 import com.bugsnag.android.Bugsnag
 import com.bugsnag.android.Configuration
+import com.bugsnag.android.createDefaultDelivery
+import com.bugsnag.android.mazerunner.InterceptingDelivery
 
 class CXXPausedSessionScenario(
     config: Configuration,
@@ -13,6 +15,10 @@ class CXXPausedSessionScenario(
     init {
         System.loadLibrary("cxx-scenarios")
         config.autoTrackSessions = false
+
+        config.delivery = InterceptingDelivery(createDefaultDelivery()) {
+            crash(0)
+        }
     }
 
     external fun crash(counter: Int): Int
@@ -21,6 +27,5 @@ class CXXPausedSessionScenario(
         super.startScenario()
         Bugsnag.startSession()
         Bugsnag.pauseSession()
-        crash(0)
     }
 }
