--- features/fixtures/mazerunner/jvm-scenarios/src/main/java/com/bugsnag/android/mazerunner/scenarios/InternalReportScenario.kt
@@ -1,47 +1,0 @@
-package com.bugsnag.android.mazerunner.scenarios
-
-import android.content.Context
-import android.os.Build
-import android.os.storage.StorageManager
-import com.bugsnag.android.Bugsnag
-import com.bugsnag.android.Configuration
-import java.io.File
-
-/**
- * Sends a handled exception to Bugsnag, which does not include session data.
- */
-internal class InternalReportScenario(
-    config: Configuration,
-    context: Context,
-    eventMetadata: String?
-) : Scenario(config, context, eventMetadata) {
-
-    init {
-        config.autoTrackSessions = false
-
-        val errDir = File(context.cacheDir, "bugsnag-errors")
-
-        if (eventMetadata == "tombstone" && Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
-            errDir.mkdir()
-            val storageManager = context.getSystemService(Context.STORAGE_SERVICE) as StorageManager
-            storageManager.setCacheBehaviorGroup(errDir, true)
-            storageManager.setCacheBehaviorTombstone(errDir, true)
-        }
-    }
-
-    override fun startBugsnag(startBugsnagOnly: Boolean) {
-        if (startBugsnagOnly) {
-            val errDir = File(context.cacheDir, "bugsnag-errors")
-            val files = errDir.listFiles()
-            files.forEach { it.writeText("{[]}") }
-        } else {
-            disableAllDelivery(config)
-        }
-    }
-
-    override fun startScenario() {
-        super.startScenario()
-
-        Bugsnag.notify(java.lang.RuntimeException("Whoops"))
-    }
-}
