--- features/full_tests/batch_1/cached_session_reports.feature
@@ -1,20 +1,0 @@
-Feature: Cached Session Reports
-
-Scenario: If an empty file is in the cache directory then zero requests should be made
-    When I run "EmptySessionScenario" and relaunch the app
-    And I configure Bugsnag for "EmptySessionScenario"
-    Then I should receive no requests
-
-# TODO: Skip pending PLAT-5488
-@skip
-Scenario: Sending internal error reports on API <26
-    When I run "PartialSessionScenario" and relaunch the app
-    And I configure Bugsnag for "PartialSessionScenario"
-    Then I should receive no requests
-
-# TODO: Skip pending PLAT-5488
-@skip
-Scenario: If a file in the cache directory is deleted before a request completes, zero further requests should be made
-    When I run "DeletedSessionScenario" and relaunch the app
-    And I configure Bugsnag for "DeletedSessionScenario"
-    Then I should receive no requests
